Update settings.py
Tweaked ALLOWED_HOSTS for additional host references that may be used for local host references.  This would still need to be changed to account for the host used for production hosting, such as 'geonews.eastus.cloudapp.azure.com' (our Azure host).  But for other testers, its best to leave as is.