Add files via upload
Bash script to make it easier to quickly export the API keys, and avoid input errors.  The keys inside can/should be replaced with the keys of the system owner.  The defaults in there now are current active keys for the CMSC 495 GeoNews project, but may not be active in the future. (7/6/2021)